Module ethcore_util::bytes

Unified interfaces for bytes operations on basic types

Examples

extern crate ethcore_util as util;

fn bytes_convertable() {
    use util::bytes::BytesConvertable;

    let arr = [0; 5];
    let slice: &[u8] = arr.bytes();
}

fn to_bytes() {
    use util::bytes::ToBytes;

    let a: Vec<u8> = "hello_world".to_bytes();
    let b: Vec<u8> = 400u32.to_bytes();
    let c: Vec<u8> = 0xffffffffffffffffu64.to_bytes();
}

fn from_bytes() {
    use util::bytes::FromBytes;

    let a = String::from_bytes(&[b'd', b'o', b'g']);
    let b = u16::from_bytes(&[0xfa]);
    let c = u64::from_bytes(&[0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff]);
}

fn main() {
    bytes_convertable();
    to_bytes();
    from_bytes();
}

Structs

PrettySlice
TODO [Gav Wood] Please document me

Enums

BytesRef
TODO [debris] Please document me

FromBytesError
Error returned when FromBytes conversation goes wrong

Traits

BytesConvertable
Slice of bytes to underlying memory

FromBytes
Converts to given type from its bytes representation

Populatable
Simple trait to allow for raw population of a Sized object from a byte slice.

ToBytes
Converts given type to its shortest representation in bytes

ToPretty
TODO [Gav Wood] Please document me

Type Definitions

Bytes
Vector of bytes

FromBytesResult
Alias for the result of FromBytes trait