Trait ethcore_util::bytes::Populatable

pub trait Populatable {
    fn as_slice_mut(&mut self) -> &mut [u8];

    fn populate_raw(&mut self, d: &[u8]) { ... }
    fn copy_raw(&mut self, d: &[u8]) { ... }
    fn populate_raw_from(&mut self, d: &BytesConvertable) { ... }
    fn copy_raw_from(&mut self, d: &BytesConvertable) { ... }
}

Simple trait to allow for raw population of a Sized object from a byte slice.

Required Methods

fn as_slice_mut(&mut self) -> &mut [u8]

Get the raw slice for this object.

Provided Methods

fn populate_raw(&mut self, d: &[u8])

Copies a bunch of bytes d to self, overwriting as necessary.

If d is smaller, zero-out the remaining bytes.

fn copy_raw(&mut self, d: &[u8])

Copies a bunch of bytes d to self, overwriting as necessary.

If d is smaller, will leave some bytes untouched.

fn populate_raw_from(&mut self, d: &BytesConvertable)

Copies the raw representation of an object d to self, overwriting as necessary.

If d is smaller, zero-out the remaining bytes.

fn copy_raw_from(&mut self, d: &BytesConvertable)

Copies the raw representation of an object d to self, overwriting as necessary.

If d is smaller, will leave some bytes untouched.