Module ethcore_util::chainfilter

Multilevel blockchain bloom filter.

extern crate ethcore_util as util;
use std::str::FromStr;
use util::chainfilter::*;
use util::sha3::*;
use util::hash::*;
 
fn main() {
    let (index_size, bloom_levels) = (16, 3);
    let mut cache = MemoryCache::new();
        
    let address = Address::from_str("ef2d6d194084c2de36e0dabfce45d046b37d1106").unwrap();
        
    // borrow cache for reading inside the scope
    let modified_blooms = {
        let filter = ChainFilter::new(&cache, index_size, bloom_levels);    
        let block_number = 39;
        let mut bloom = H2048::new();
        bloom.shift_bloomed(&address.sha3());
        filter.add_bloom(&bloom, block_number)
    };
        
    // number of updated blooms is equal number of levels
    assert_eq!(modified_blooms.len(), bloom_levels as usize);

    // lets inserts modified blooms into the cache
    cache.insert_blooms(modified_blooms);
        
    // borrow cache for another reading operations
    {
        let filter = ChainFilter::new(&cache, index_size, bloom_levels);    
        let blocks = filter.blocks_with_address(&address, 10, 40);
        assert_eq!(blocks.len(), 1);    
        assert_eq!(blocks[0], 39);
    }
}

Structs

BloomIndex
Represents bloom index in cache

ChainFilter
Should be used for search operations on blockchain.

MemoryCache
In memory cache for blooms.

Traits

FilterDataSource
Types implementing this trait should provide read access for bloom filters database.