Struct ethcore_util::chainfilter::BloomIndex

pub struct BloomIndex {
    pub level: u8,
    pub index: usize,
}

Represents bloom index in cache

On cache level 0, every block bloom is represented by different index. On higher cache levels, multiple block blooms are represented by one index. Their BloomIndex can be created from block number and given level.

Fields

level

TODO [debris] Please document me

index

TODO [debris] Please document me

Methods

impl BloomIndex

fn new(level: u8, index: usize) -> BloomIndex

Default constructor for BloomIndex