Enum ethcore_util::error::UtilError

pub enum UtilError {
    StdIo(Error),
    FromHex(FromHexError),
    BaseData(BaseDataError),
    Decoder(DecoderError),
    SimpleString(String),
    BadSize,
    Snappy(InvalidInput),
}

General error type which should be capable of representing all errors in ethcore.

Variants
StdIo(Error)

Error concerning the Rust standard library's IO subsystem.

FromHex(FromHexError)

Error concerning the hex conversion logic.

BaseData(BaseDataError)

Error concerning the database abstraction logic.

Decoder(DecoderError)

Error concerning the RLP decoder.

SimpleString(String)

Miscellaneous error described by a string.

BadSize

Error from a bad input size being given for the needed output.

Snappy(InvalidInput)

Error from snappy.