Module ethcore_util::hash

General hash types, a fixed-size raw-data type used as the output of hash functions.

Structs

Address
TODO [Gav Wood] Please document me

H1024
TODO [Gav Wood] Please document me

H128
TODO [Gav Wood] Please document me

H2048
TODO [Gav Wood] Please document me

H256
TODO [Gav Wood] Please document me

H264
TODO [Gav Wood] Please document me

H32
TODO [Gav Wood] Please document me

H512
TODO [Gav Wood] Please document me

H520
TODO [Gav Wood] Please document me

H64
TODO [Gav Wood] Please document me

Statics

ZERO_ADDRESS
Constant address for point 0. Often used as a default.

ZERO_H256
Constant 256-bit datum for 0. Often used as a default.

Traits

FixedHash
Trait for a fixed-size byte array to be used as the output of hash functions.

Functions

address_from_hex
TODO [Gav Wood] Please document me

address_from_u64
TODO [Gav Wood] Please document me

h256_from_hex
TODO [Gav Wood] Please document me

h256_from_u64
TODO [Gav Wood] Please document me