Trait ethcore_util::io::IoHandler

pub trait IoHandler<Message>: Send where Message: Send + 'static {
    fn initialize<'s>(&'s mut self, _io: &mut IoContext<'s, Message>) { ... }
    fn timeout<'s>(&'s mut self, _io: &mut IoContext<'s, Message>, _timer: TimerToken) { ... }
    fn message<'s>(&'s mut self, _io: &mut IoContext<'s, Message>, _message: &'s mut Message) { ... }
    fn stream_hup<'s>(&'s mut self, _io: &mut IoContext<'s, Message>, _stream: StreamToken) { ... }
    fn stream_readable<'s>(&'s mut self, _io: &mut IoContext<'s, Message>, _stream: StreamToken) { ... }
    fn stream_writable<'s>(&'s mut self, _io: &mut IoContext<'s, Message>, _stream: StreamToken) { ... }
}

Generic IO handler. All the handler function are called from within IO event loop. Message type is used as notification data

Provided Methods

fn initialize<'s>(&'s mut self, _io: &mut IoContext<'s, Message>)

Initialize the handler

fn timeout<'s>(&'s mut self, _io: &mut IoContext<'s, Message>, _timer: TimerToken)

Timer function called after a timeout created with HandlerIo::timeout.

fn message<'s>(&'s mut self, _io: &mut IoContext<'s, Message>, _message: &'s mut Message)

Called when a broadcasted message is received. The message can only be sent from a different IO handler.

fn stream_hup<'s>(&'s mut self, _io: &mut IoContext<'s, Message>, _stream: StreamToken)

Called when an IO stream gets closed

fn stream_readable<'s>(&'s mut self, _io: &mut IoContext<'s, Message>, _stream: StreamToken)

Called when an IO stream can be read from

fn stream_writable<'s>(&'s mut self, _io: &mut IoContext<'s, Message>, _stream: StreamToken)

Called when an IO stream can be written to