Struct ethcore_util::journaldb::JournalDB

pub struct JournalDB {
    // some fields omitted
}

Implementation of the HashDB trait for a disk-backed database with a memory overlay and latent-removal semantics.

Like OverlayDB, there is a memory overlay; commit() must be called in order to write operations out to disk. Unlike OverlayDB, remove() operations do not take effect immediately. Rather some age (based on a linear but arbitrary metric) must pass before the removals actually take effect.

Methods

impl JournalDB

fn new(backing: DB) -> JournalDB

Create a new instance given a backing database.

fn new_temp() -> JournalDB

Create a new instance with an anonymous temporary database.

fn to_overlaydb(&self) -> OverlayDB

Get a clone of the overlay db portion of this.

fn commit(&mut self, now: u64, id: &H256, end: Option<(u64, H256)>) -> Result<u32, UtilError>

Commit all recent insert operations and historical removals from the old era to the backing database.

fn revert(&mut self)

Revert all operations on this object (i.e. insert()s and removes()s) since the last commit().