Trait ethcore_util::journaldb::traits::JournalDB

pub trait JournalDB: HashDB + Send + Sync {
    fn boxed_clone(&self) -> Box<JournalDB>;
    fn mem_used(&self) -> usize;
    fn is_empty(&self) -> bool;
    fn latest_era(&self) -> Option<u64>;
    fn commit(&mut self, now: u64, id: &H256, end: Option<(u64, H256)>) -> Result<u32, UtilError>;

    fn state(&self, _id: &H256) -> Option<Bytes> { ... }
    fn is_pruned(&self) -> bool { ... }
}

A HashDB which can manage a short-term journal potentially containing many forks of mutually exclusive actions.

Required Methods

fn boxed_clone(&self) -> Box<JournalDB>

Return a copy of ourself, in a box.

fn mem_used(&self) -> usize

Returns heap memory size used

fn is_empty(&self) -> bool

Check if this database has any commits

fn latest_era(&self) -> Option<u64>

Get the latest era in the DB. None if there isn't yet any data in there.

fn commit(&mut self, now: u64, id: &H256, end: Option<(u64, H256)>) -> Result<u32, UtilError>

Commit all recent insert operations and canonical historical commits' removals from the old era to the backing database, reverting any non-canonical historical commit's inserts.

Provided Methods

fn state(&self, _id: &H256) -> Option<Bytes>

State data query

fn is_pruned(&self) -> bool

Whether this database is pruned.