Struct ethcore_util::keys::directory::KeyDirectory

pub struct KeyDirectory {
    // some fields omitted
}

Represents directory for saving/loading key files.

Methods

impl KeyDirectory

fn new(path: &Path) -> KeyDirectory

Initializes new cache directory context with a given path

fn save(&mut self, key_file: KeyFileContent) -> Result<Uuid, Error>

saves (inserts or updates) given key

fn get(&self, id: &Uuid) -> Option<KeyFileContent>

Returns key given by id if corresponding file exists and no load error occured. Warns if any error occured during the key loading

fn path(&self) -> &str

Returns current path to the directory with keys

fn collect_garbage(&mut self)

Removes keys that never been requested during last MAX_USAGE_TRACK times

fn cache_size(&self) -> usize

Reports how many keys are currently cached.

fn delete(&mut self, id: &Uuid) -> Result<(), Error>

Removes key file from key directory

fn list(&self) -> Result<Vec<Uuid>, Error>

Enumerates all keys in the directory