Struct ethcore_util::keys::directory::KeyFileContent

pub struct KeyFileContent {
    pub crypto: KeyFileCrypto,
    pub id: Uuid,
    pub account: Option<Address>,
    // some fields omitted
}

Stored key file struct with encrypted message (cipher_text) also contains password derivation function settings (PBKDF2/Scrypt)

Fields

crypto

Holds cypher and decrypt function settings.

id

The identifier.

account

Account (if present)

Methods

impl KeyFileContent

fn new(crypto: KeyFileCrypto) -> KeyFileContent

New stored key file struct with encrypted message (cipher_text) also contains password derivation function settings (PBKDF2/Scrypt) to decrypt cipher_text given the password is provided.

fn load(json: &Json) -> Result<KeyFileContent, ()>

Loads key from valid json, returns error and records warning if key is mallformed

fn version(&self) -> Option<u64>

Returns key file version if it is known.