Struct ethcore_util::keys::store::SecretStore

pub struct SecretStore {
    // some fields omitted
}

Represent service for storing encrypted arbitrary data

Methods

impl SecretStore

fn new_in(path: &Path) -> Self

new instance of Secret Store in specific directory

fn try_import_existing(&mut self)

trys to import keys in the known locations

fn accounts(&self) -> Result<Vec<(Address, H128)>, Error>

Lists all accounts and corresponding key ids

fn account(&self, account: &Address) -> Option<H128>

Resolves key_id by account address

fn import_key(&mut self, key_file: KeyFileContent) -> Result<(), Error>

Imports pregenerated key, returns error if not saved correctly

fn unlock_account(&self, account: &Address, pass: &str) -> Result<(), EncryptedHashMapError>

Unlocks account for use

fn unlock_account_no_expire(&self, account: &Address, pass: &str) -> Result<(), EncryptedHashMapError>

Unlocks account for use (no expiration of unlock)

fn new_account(&mut self, pass: &str) -> Result<Address, Error>

Creates new account

fn sign(&self, account: &Address, message: &H256) -> Result<Signature, SigningError>

Signs message with unlocked account

fn account_secret(&self, account: &Address) -> Result<Secret, SigningError>

Returns secret for unlocked account

fn collect_garbage(&mut self)

Makes account unlocks expire and removes unused key files from memory