Module ethcore_util::log

Common log helper functions

Structs

RotatingLogger
Logger implementation that keeps up to LOG_SIZE log elements.

Style
A style is a collection of properties that can format a string using ANSI escape codes.

Enums

Colour
A colour is one specific type of ANSI escape code, and can refer to either the foreground or background colour.

Functions

init_log
Intialize log with default settings