Struct ethcore_util::memorydb::MemoryDB

pub struct MemoryDB {
    // some fields omitted
}

Reference-counted memory-based HashDB implementation.

Use new() to create a new database. Insert items with insert(), remove items with kill(), check for existance with exists() and lookup a hash to derive the data with lookup(). Clear with clear() and purge the portions of the data that have no references with purge().

Example

extern crate ethcore_util;
use ethcore_util::hashdb::*;
use ethcore_util::memorydb::*;
fn main() {
  let mut m = MemoryDB::new();
  let d = "Hello world!".as_bytes();

  let k = m.insert(d);
  assert!(m.exists(&k));
  assert_eq!(m.lookup(&k).unwrap(), d);

  m.insert(d);
  assert!(m.exists(&k));

  m.kill(&k);
  assert!(m.exists(&k));

  m.kill(&k);
  assert!(!m.exists(&k));

  m.kill(&k);
  assert!(!m.exists(&k));

  m.insert(d);
  assert!(!m.exists(&k));
  m.insert(d);
  assert!(m.exists(&k));
  assert_eq!(m.lookup(&k).unwrap(), d);

  m.kill(&k);
  assert!(!m.exists(&k));
}

Methods

impl MemoryDB

fn new() -> MemoryDB

Create a new instance of the memory DB.

fn clear(&mut self)

Clear all data from the database.

Examples

extern crate ethcore_util;
use ethcore_util::hashdb::*;
use ethcore_util::memorydb::*;
fn main() {
  let mut m = MemoryDB::new();
  let hello_bytes = "Hello world!".as_bytes();
  let hash = m.insert(hello_bytes);
  assert!(m.exists(&hash));
  m.clear();
  assert!(!m.exists(&hash));
}

fn purge(&mut self)

Purge all zero-referenced data from the database.

fn raw(&self, key: &H256) -> Option<&(Bytes, i32)>

Grab the raw information associated with a key. Returns None if the key doesn't exist.

Even when Some is returned, the data is only guaranteed to be useful when the refs > 0.

fn drain(&mut self) -> HashMap<H256, (Bytes, i32)>

TODO [Gav Wood] Please document me

fn denote(&self, key: &H256, value: Bytes) -> &(Bytes, i32)

TODO [Gav Wood] Please document me