Struct ethcore_util::migration::Batch

pub struct Batch { /* fields omitted */ }

A batch of key-value pairs to be written into the database.

Methods

impl Batch

fn new(config: &Config, col: Option<u32>) -> Self

Make a new batch with the given config.

fn insert(&mut self,
key: Vec<u8>,
value: Vec<u8>,
dest: &mut Database)
-> Result<(), Error>

Insert a value into the batch, committing if necessary.

fn commit(&mut self, dest: &mut Database) -> Result<(), Error>

Commit all the items in the batch to the given database.