Trait ethcore_util::migration::Migration

pub trait Migration: 'static {
    fn columns(&self) -> Option<u32>;
    fn version(&self) -> u32;
    fn migrate(&mut self,
           source: Arc<Database>,
           config: &Config,
           destination: &mut Database,
           col: Option<u32>)
           -> Result<(), Error>;

    fn pre_columns(&self) -> Option<u32> { ... }
}

A generalized migration from the given db to a destination db.

Required Methods

fn columns(&self) -> Option<u32>

Number of columns in database after the migration.

fn version(&self) -> u32

Version of the database after the migration.

fn migrate(&mut self,
source: Arc<Database>,
config: &Config,
destination: &mut Database,
col: Option<u32>)
-> Result<(), Error>

Migrate a source to a destination.

Provided Methods

fn pre_columns(&self) -> Option<u32>

Number of columns in the database before the migration.