Module ethcore_util::network

TODO [Gav Wood] Please document me

Reexports

pub use network::host::NetworkIoMessage::User as UserMessage;

Traits

NetworkProtocolHandler
Network IO protocol handler. This needs to be implemented for each new subprotocol. All the handler function are called from within IO event loop. Message is the type for message data.

Type Definitions

NetworkContext
TODO [arkpar] Please document me

NetworkError
TODO [arkpar] Please document me

NetworkIoMessage
TODO [arkpar] Please document me

NetworkService
TODO [arkpar] Please document me

PacketId
TODO [arkpar] Please document me

PeerId
TODO [arkpar] Please document me