Struct ethcore_util::network::NetworkService

pub struct NetworkService<Message> where Message: Send + Sync + Clone + 'static {
    // some fields omitted
}

IO Service with networking Message defines a notification data type.

Methods

impl<Message> NetworkService<Message> where Message: Send + Sync + Clone + 'static

fn new(config: NetworkConfiguration) -> Result<NetworkService<Message>, UtilError>

Starts IO event loop

fn register_protocol(&self, handler: Arc<NetworkProtocolHandler<Message> + Send + Sync>, protocol: ProtocolId, versions: &[u8]) -> Result<(), NetworkError>

Regiter a new protocol handler with the event loop.

fn host_info(&self) -> String

Returns host identifier string as advertised to other peers

fn io(&self) -> &IoService<NetworkIoMessage<Message>>

Returns underlying io service.

fn stats(&self) -> &NetworkStats

Returns network statistics.

fn external_url(&self) -> Option<String>

Returns external url if available.

fn local_url(&self) -> Option<String>

Returns external url if available.

fn start(&self) -> Result<(), UtilError>

Start network IO

fn stop(&self) -> Result<(), UtilError>

Stop network IO

fn add_reserved_peer(&self, peer: &str) -> Result<(), UtilError>

Try to add a reserved peer.

fn remove_reserved_peer(&self, peer: &str) -> Result<(), UtilError>

Try to remove a reserved peer.

fn set_non_reserved_mode(&self, mode: NonReservedPeerMode)

Set the non-reserved peer mode.