Struct ethcore_util::overlaydb::OverlayDB

pub struct OverlayDB {
    // some fields omitted
}

Implementation of the HashDB trait for a disk-backed database with a memory overlay.

The operations insert() and remove() take place on the memory overlay; batches of such operations may be flushed to the disk-backed DB with commit() or discarded with revert().

lookup() and contains() maintain normal behaviour - all insert() and remove() queries have an immediate effect in terms of these functions.

Methods

impl OverlayDB

fn new(backing: DB) -> OverlayDB

Create a new instance of OverlayDB given a backing database.

fn new_with_arc(backing: Arc<DB>) -> OverlayDB

Create a new instance of OverlayDB given a backing database.

fn new_temp() -> OverlayDB

Create a new instance of OverlayDB with an anonymous temporary database.

fn commit(&mut self) -> Result<u32, UtilError>

Commit all memory operations to the backing database.

Returns either an error or the number of items changed in the backing database.

Will return an error if the number of kill()s ever exceeds the number of insert()s for any key. This will leave the database in an undeterminate state. Don't ever let it happen.

Example

extern crate ethcore_util;
use ethcore_util::hashdb::*;
use ethcore_util::overlaydb::*;
fn main() {
  let mut m = OverlayDB::new_temp();
  let key = m.insert(b"foo");           // insert item.
  assert!(m.exists(&key));              // key exists (in memory).
  assert_eq!(m.commit().unwrap(), 1);   // 1 item changed.
  assert!(m.exists(&key));              // key still exists (in backing).
  m.kill(&key);                         // delete item.
  assert!(!m.exists(&key));             // key "doesn't exist" (though still does in backing).
  m.kill(&key);                         // oh dear... more kills than inserts for the key...
  //m.commit().unwrap();                // this commit/unwrap would cause a panic.
  m.revert();                           // revert both kills.
  assert!(m.exists(&key));              // key now still exists.
}

fn revert(&mut self)

Revert all operations on this object (i.e. insert()s and kill()s) since the last commit().

Example

extern crate ethcore_util;
use ethcore_util::hashdb::*;
use ethcore_util::overlaydb::*;
fn main() {
  let mut m = OverlayDB::new_temp();
  let foo = m.insert(b"foo");   // insert foo.
  m.commit().unwrap();          // commit - new operations begin here...
  let bar = m.insert(b"bar");   // insert bar.
  m.kill(&foo);                 // kill foo.
  assert!(!m.exists(&foo));     // foo is gone.
  assert!(m.exists(&bar));      // bar is here.
  m.revert();                   // revert the last two operations.
  assert!(m.exists(&foo));      // foo is here.
  assert!(!m.exists(&bar));     // bar is gone.
}