Module ethcore_util::panics

Panic utilities

Structs

PanicHandler
Structure that allows to catch panics and notify listeners

Traits

ForwardPanic
Forwards panics from child

MayPanic
Trait indicating that the structure catches some of the panics (most probably from spawned threads) and it's possbile to be notified when one of the threads panics.

OnPanicListener
Thread-safe closure for handling possible panics