Module ethcore_util::rlp

Rlp serialization module

Allows encoding, decoding, and view onto rlp-slice

What should you use when?

Use encode function when:

You want to encode something inline.

You do not work on big set of data.

You want to encode whole data structure at once.

Use decode function when:

You want to decode something inline.

You do not work on big set of data.

You want to decode whole rlp at once.

Use RlpStream when:

You want to encode something in portions.

You encode a big set of data.

Use Rlp when:

You are working on trusted data (not corrupted).

You want to get view onto rlp-slice.

You don't want to decode whole rlp at once.

Use UntrustedRlp when:

You are working on untrusted data (~corrupted).

You need to handle data corruption errors.

You are working on input data.

You want to get view onto rlp-slice.

You don't want to decode whole rlp at once.

Reexports

pub use self::rlperrors::DecoderError;

pub use self::rlptraits::{Decoder, Decodable, View, Stream, Encodable, Encoder};

pub use self::untrusted_rlp::{UntrustedRlp, UntrustedRlpIterator, PayloadInfo, Prototype};

pub use self::rlpin::{Rlp, RlpIterator};

pub use self::rlpstream::{RlpStream, RlpStandard};

Modules

rlperrors
TODO [Gav Wood] Please document me

rlpin
TODO [debris] Please document me

rlpstream
TODO [debris] Please document me

rlptraits
TODO [Gav Wood] Please document me

untrusted_rlp
TODO [debris] Please document me

Constants

EMPTY_LIST_RLP
TODO [Gav Wood] Please document me

NULL_RLP
TODO [arkpar] Please document me

SHA3_EMPTY_LIST_RLP
TODO [debris] Please document me

SHA3_NULL_RLP
TODO [arkpar] Please document me

Functions

decode
Shortcut function to decode trusted rlp

encode
Shortcut function to encode structure into rlp.