Module ethcore_util::squeeze

Helper module that should be used to randomly squeeze caches to a given size in bytes

extern crate heapsize;
extern crate ethcore_util as util;
use std::collections::HashMap;
use std::mem::size_of;
use heapsize::HeapSizeOf;
use util::squeeze::Squeeze;
 
fn main() {
    let initial_size = 60;
    let mut map: HashMap<u8, u8> = HashMap::with_capacity(initial_size);
    assert!(map.capacity() >= initial_size);
    for i in 0..initial_size {
        map.insert(i as u8, i as u8);
    }
    
    assert_eq!(map.heap_size_of_children(), map.capacity() * 2 * size_of::<u8>());
    assert_eq!(map.len(), initial_size);
    let initial_heap_size = map.heap_size_of_children();

    // squeeze it to size of key and value
    map.squeeze(2 * size_of::<u8>());
    assert_eq!(map.len(), 1);
    
    // its likely that heap size was reduced, but we can't be 100% sure
    assert!(initial_heap_size >= map.heap_size_of_children());
}

Traits

Squeeze
Should be used to squeeze collections to certain size in bytes