Module ethcore_util::standard

TODO [Gav Wood] Please document me

Reexports

pub use std::sync::*;

pub use std::ops::*;

pub use std::cmp::*;

pub use std::cell::*;

pub use std::collections::*;

Modules

cmp
Functionality for ordering and comparison.

fmt
Utilities for formatting and printing strings

io
Traits, helpers, and type definitions for core I/O functionality.

mem
Basic functions for dealing with memory

ops
Overloadable operators

option
Optional values

ptr
Raw, unsafe pointers, *const T, and *mut T

result
Error handling with the Result type

slice
A dynamically-sized view into a contiguous sequence, [T].

str
Unicode string slices

Structs

Path
A slice of a path (akin to str).

Enums

FromHexError
Errors that can occur when decoding a hex encoded string

Json
Represents a json value

Traits

FromBase64
A trait for converting from base64 encoded values.

FromHex
A trait for converting hexadecimal encoded values

FromStr
A trait to abstract the idea of creating a new instance of a type from a string.

Hash
A hashable type.

Hasher
A trait which represents the ability to hash an arbitrary stream of bytes.

HeapSizeOf

Itertools
The trait Itertools: extra iterator adaptors and methods for iterators.

Read
The Read trait allows for reading bytes from a source.

StdError
Base functionality for all errors in Rust.

Write
A trait for objects which are byte-oriented sinks.