Enum ethcore_util::standard::Json

pub enum Json {
    I64(i64),
    U64(u64),
    F64(f64),
    String(String),
    Boolean(bool),
    Array(Vec<Json>),
    Object(BTreeMap<String, Json>),
    Null,
}

Represents a json value

Variants

I64

U64

F64

String

Boolean

Array

Object

Null

Methods

impl Json

fn from_reader(rdr: &mut Read) -> Result<Json, ParserError>

Decodes a json value from an &mut io::Read

fn from_str(s: &str) -> Result<Json, ParserError>

Decodes a json value from a string

fn pretty(&self) -> PrettyJson

Borrow this json object as a pretty object to generate a pretty representation for it via Display.

fn find(&'a self, key: &str) -> Option<&'a Json>

If the Json value is an Object, returns the value associated with the provided key. Otherwise, returns None.

fn find_path(&'a self, keys: &[&str]) -> Option<&'a Json>

Attempts to get a nested Json Object for each key in keys. If any key is found not to exist, find_path will return None. Otherwise, it will return the Json value associated with the final key.

fn search(&'a self, key: &str) -> Option<&'a Json>

If the Json value is an Object, performs a depth-first search until a value associated with the provided key is found. If no value is found or the Json value is not an Object, returns None.

fn is_object(&'a self) -> bool

Returns true if the Json value is an Object. Returns false otherwise.

fn as_object(&'a self) -> Option<&'a BTreeMap<String, Json>>

If the Json value is an Object, returns the associated BTreeMap. Returns None otherwise.

fn as_object_mut(&'a mut self) -> Option<&'a mut BTreeMap<String, Json>>

If the Json value is an Object, returns the associated mutable BTreeMap. Returns None otherwise.

fn is_array(&'a self) -> bool

Returns true if the Json value is an Array. Returns false otherwise.

fn as_array(&'a self) -> Option<&'a Vec<Json>>

If the Json value is an Array, returns the associated vector. Returns None otherwise.

fn as_array_mut(&'a mut self) -> Option<&'a mut Vec<Json>>

If the Json value is an Array, returns the associated mutable vector. Returns None otherwise.

fn is_string(&'a self) -> bool

Returns true if the Json value is a String. Returns false otherwise.

fn as_string(&'a self) -> Option<&'a str>

If the Json value is a String, returns the associated str. Returns None otherwise.

fn is_number(&self) -> bool

Returns true if the Json value is a Number. Returns false otherwise.

fn is_i64(&self) -> bool

Returns true if the Json value is a i64. Returns false otherwise.

fn is_u64(&self) -> bool

Returns true if the Json value is a u64. Returns false otherwise.

fn is_f64(&self) -> bool

Returns true if the Json value is a f64. Returns false otherwise.

fn as_i64(&self) -> Option<i64>

If the Json value is a number, return or cast it to a i64. Returns None otherwise.

fn as_u64(&self) -> Option<u64>

If the Json value is a number, return or cast it to a u64. Returns None otherwise.

fn as_f64(&self) -> Option<f64>

If the Json value is a number, return or cast it to a f64. Returns None otherwise.

fn is_boolean(&self) -> bool

Returns true if the Json value is a Boolean. Returns false otherwise.

fn as_boolean(&self) -> Option<bool>

If the Json value is a Boolean, returns the associated bool. Returns None otherwise.

fn is_null(&self) -> bool

Returns true if the Json value is a Null. Returns false otherwise.

fn as_null(&self) -> Option<()>

If the Json value is a Null, returns (). Returns None otherwise.