Struct ethcore_util::standard::io::Stdout

pub struct Stdout { /* fields omitted */ }

A handle to the global standard output stream of the current process.

Each handle shares a global buffer of data to be written to the standard output stream. Access is also synchronized via a lock and explicit control over locking is available via the lock method.

Created by the io::stdout method.

Methods

impl Stdout

fn lock(&self) -> StdoutLock

Locks this handle to the standard output stream, returning a writable guard.

The lock is released when the returned lock goes out of scope. The returned guard also implements the Write trait for writing data.

Examples

use std::io::{self, Write};

let stdout = io::stdout();
let mut handle = stdout.lock();

try!(handle.write(b"hello world"));