Struct ethcore_util::standard::ops::RangeTo

pub struct RangeTo<Idx> {
    pub end: Idx,
}

A range which is only bounded above: { x | x < end }. Use ..end (two dots) for its shorthand.

See the contains() method for its characterization.

It cannot serve as an iterator because it doesn't have a starting point.

Examples

The ..{integer} syntax is a RangeTo:

assert_eq!((..5), std::ops::RangeTo{ end: 5 });

It does not have an IntoIterator implementation, so you can't use it in a for loop directly. This won't compile:

for i in ..5 {
    // ...
}

When used as a slicing index, RangeTo produces a slice of all array elements before the index indicated by end.

let arr = [0, 1, 2, 3];
assert_eq!(arr[ .. ], [0,1,2,3]);
assert_eq!(arr[ ..3], [0,1,2  ]);  // RangeTo
assert_eq!(arr[1.. ], [  1,2,3]);
assert_eq!(arr[1..3], [  1,2  ]);

Fields
end: Idx

The upper bound of the range (exclusive).

Methods

impl<Idx> RangeTo<Idx> where Idx: PartialOrd<Idx>

fn contains(&self, item: Idx) -> bool

Unstable (range_contains)
: recently added as per RFC

Examples

#![feature(range_contains)]
fn main() {
    assert!(   (..5).contains(-1_000_000_000));
    assert!(   (..5).contains(4));
    assert!( ! (..5).contains(5));
}