Trait ethcore_util::standard::HeapSizeOf

pub trait HeapSizeOf {
    fn heap_size_of_children(&self) -> usize;
}

Required Methods

fn heap_size_of_children(&self) -> usize

Measure the size of any heap-allocated structures that hang off this value, but not the space taken up by the value itself (i.e. what size_of:: measures, more or less); that space is handled by the implementation of HeapSizeOf for Box below.