Struct ethcore_util::table::Table

pub struct Table<Row, Col, Val> where Row: Eq + Hash + Clone, Col: Eq + Hash {
    // some fields omitted
}

Structure to hold double-indexed values

You can obviously use HashMap<(Row,Col), Val>, but this structure gives you better access to all Columns in Specific Row. Namely you can get sub-hashmap HashMap<Col, Val> for specific Row

Methods

impl<Row, Col, Val> Table<Row, Col, Val> where Row: Eq + Hash + Clone, Col: Eq + Hash

fn new() -> Self

Creates new Table

fn clear(&mut self)

Removes all elements from this Table

fn len(&self) -> usize

Returns length of the Table (number of (row, col, val) tuples)

fn is_empty(&self) -> bool

Check if there is any element in this Table

fn row_mut(&mut self, row: &Row) -> Option<&mut HashMap<Col, Val>>

Get mutable reference for single Table row.

fn has_row(&self, row: &Row) -> bool

Checks if row is defined for that table (note that even if defined it might be empty)

fn row(&self, row: &Row) -> Option<&HashMap<Col, Val>>

Get immutable reference for single row in this Table

fn get(&self, row: &Row, col: &Col) -> Option<&Val>

Get element in cell described by (row, col)

fn remove(&mut self, row: &Row, col: &Col) -> Option<Val>

Remove value from specific cell

It will remove the row if it's the last value in it

fn clear_if_empty(&mut self, row: &Row)

Remove given row from Table if there are no values defined in it

When using #row_mut it may happen that all values from some row are drained. Table however will not be aware that row is empty. You can use this method to explicitly remove row entry from the Table.

fn insert(&mut self, row: Row, col: Col, val: Val) -> Option<Val>

Inserts new value to specified cell

Returns previous value (if any)