Struct ethcore_util::trie::journal::Journal

pub struct Journal(_);

A journal of operations on the backing database.

Methods

impl Journal

fn new() -> Journal

Create a new, empty, object.

fn new_node(&mut self, rlp: Bytes, out: &mut RlpStream)

Given the RLP that encodes a node, append a reference to that node out and leave journal such that the reference is valid, once applied.

fn delete_node_sha3(&mut self, old_sha3: H256)

Given the RLP that encodes a now-unused node, leave journal in such a state that it is noted.

fn delete_node(&mut self, old: &[u8])

Register an RLP-encoded node for deletion (given a slice), if it needs to be deleted.

fn apply(self, db: &mut HashDB) -> Score

Apply this journal to the HashDB db and return the number of insertions and removals done.