Enum ethcore_util::trie::node::Node

pub enum Node<'a> {
    Empty,
    Leaf(NibbleSlice<'a>, &'a [u8]),
    Extension(NibbleSlice<'a>, &'a [u8]),
    Branch([&'a [u8]; 16], Option<&'a [u8]>),
}

Type of node in the trie and essential information thereof.

Variants

Empty

TODO [Gav Wood] Please document me

Leaf

TODO [Gav Wood] Please document me

Extension

TODO [Gav Wood] Please document me

Branch

TODO [Gav Wood] Please document me

Methods

impl<'a> Node<'a>

fn decoded(node_rlp: &'a [u8]) -> Node<'a>

Decode the node_rlp and return the Node.

fn encoded(&self) -> Bytes

Encode the node into RLP.

Will always return the direct node RLP even if it's 32 or more bytes. To get the RLP which would be valid for using in another node, use encoded_and_added().

fn encoded_and_added(&self, journal: &mut Journal) -> Bytes

Encode the node, adding it to journal if necessary and return the RLP valid for insertion into a parent node.