Struct ethcore_util::trie::recorder::BasicRecorder

pub struct BasicRecorder { /* fields omitted */ }

A simple recorder. Does nothing fancy but fulfills the Recorder interface properly.

Methods

impl BasicRecorder

fn new() -> Self

Create a new BasicRecorder which records all given nodes.

fn with_depth(depth: u32) -> Self

Create a BasicRecorder which only records nodes beyond a given depth.