Trait ethcore_util::trie::recorder::Recorder

pub trait Recorder {
    fn record(&mut self, hash: &H256, data: &[u8], depth: u32);
    fn drain(&mut self) -> Vec<Record> where Self: Sized;
}

Trie node recorder.

These are used to record which nodes are visited during a trie query. Inline nodes are not to be recorded, as they are contained within their parent.

Required Methods

fn record(&mut self, hash: &H256, data: &[u8], depth: u32)

Record that the given node has been visited.

The depth parameter is the depth of the visited node, with the root node having depth 0.

fn drain(&mut self) -> Vec<Record> where Self: Sized

Drain all accepted records from the recorder in ascending order by depth.