Struct ethcore_util::trie::sectriedbmut::SecTrieDBMut

pub struct SecTrieDBMut<'db> {
    // some fields omitted
}

A mutable Trie implementation which hashes keys and uses a generic HashDB backing database.

Use it as a Trie or TrieMut trait object. You can use raw() to get the backing TrieDBMut object.

Methods

impl<'db> SecTrieDBMut<'db>

fn new(db: &'db mut HashDB, root: &'db mut H256) -> Self

Create a new trie with the backing database db and empty root Initialise to the state entailed by the genesis block. This guarantees the trie is built correctly.

fn from_existing(db: &'db mut HashDB, root: &'db mut H256) -> Self

Create a new trie with the backing database db and root Panics, if root does not exist