Enum ethcore_util::trie::standardmap::Alphabet

pub enum Alphabet {
    All,
    Low,
    Mid,
    Custom(Bytes),
}

Alphabet to use when creating words for insertion into tries.

Variants
All

All values are allowed in each bytes of the key.

Low

Only a 6 values ('a' - 'f') are chosen to compose the key.

Mid

Quite a few values (around 32) are chosen to compose the key.

Custom(Bytes)

A set of bytes given is used to compose the key.