Struct ethcore_util::trie::FatDBMut

pub struct FatDBMut<'db> { /* fields omitted */ }

A mutable Trie implementation which hashes keys and uses a generic HashDB backing database. Additionaly it stores inserted hash-key mappings for later retrieval.

Use it as a Trie or TrieMut trait object.

Methods

impl<'db> FatDBMut<'db>

fn new(db: &'db mut HashDB, root: &'db mut H256) -> Self

Create a new trie with the backing database db and empty root Initialise to the state entailed by the genesis block. This guarantees the trie is built correctly.

fn from_existing(db: &'db mut HashDB, root: &'db mut H256) -> Result<Self>

Create a new trie with the backing database db and root.

Returns an error if root does not exist.

fn db(&self) -> &HashDB

Get the backing database.

fn db_mut(&mut self) -> &mut HashDB

Get the backing database.