Struct ethcore_util::trie::triedb::TrieDB

pub struct TrieDB<'db> {
    pub hash_count: usize,
    // some fields omitted
}

A Trie implementation using a generic HashDB backing database.

Use it as a Trie trait object. You can use db() to get the backing database object, keys to get the keys belonging to the trie in the backing database, and db_items_remaining() to get which items in the backing database do not belong to this trie. If this is the only trie in the backing database, then db_items_remaining() should be empty.

Example

extern crate ethcore_util as util;
use util::trie::*;
use util::hashdb::*;
use util::memorydb::*;
use util::hash::*;
use util::rlp::*;

fn main() {
  let mut memdb = MemoryDB::new();
  let mut root = H256::new();
  TrieDBMut::new(&mut memdb, &mut root).insert(b"foo", b"bar");
  let t = TrieDB::new(&memdb, &root);
  assert!(t.contains(b"foo"));
  assert_eq!(t.get(b"foo").unwrap(), b"bar");
  assert!(t.db_items_remaining().is_empty());
}

Fields

hash_count

TODO [Gav Wood] Please document me

Methods

impl<'db> TrieDB<'db>

fn new(db: &'db HashDB, root: &'db H256) -> Self

Create a new trie with the backing database db and root Panics, if root does not exist

fn db(&'db self) -> &'db HashDB

Get the backing database.

fn keys(&self) -> Vec<H256>

Determine all the keys in the backing database that belong to the trie.

fn to_map(hashes: Vec<H256>) -> HashMap<H256, u32>

Convert a vector of hashes to a hashmap of hash to occurances.

fn db_items_remaining(&self) -> HashMap<H256, i32>

Determine occurances of items in the backing database which are not related to this trie.

impl<'db> TrieDB<'db>

fn iter(&self) -> TrieDBIterator

Get all keys/values stored in the trie.