Struct ethcore_util::using_queue::UsingQueue

pub struct UsingQueue<T> where T: Clone {
    // some fields omitted
}

Special queue-like datastructure that includes the notion of usage to avoid items that were queued but never used from making it into the queue.

Methods

impl<T> UsingQueue<T> where T: Clone

fn new(max_size: usize) -> UsingQueue<T>

Create a new struct with a maximum size of max_size.

fn peek_last_ref(&self) -> Option<&T>

Return a reference to the item at the top of the queue (or None if the queue is empty); it doesn't constitute noting that the item is used.

fn use_last_ref(&mut self) -> Option<&T>

Return a reference to the item at the top of the queue (or None if the queue is empty); this constitutes using the item and will remain in the queue for at least another max_size invocations of push().

fn push(&mut self, b: T)

Place an item on the end of the queue. The previously push()ed item will be removed if use_last_ref() since it was push()ed.

fn reset(&mut self)

Clears everything; the queue is entirely reset.

fn take_used_if<P>(&mut self, predicate: P) -> Option<T> where P: Fn(&T) -> bool

Returns Some item which is the first that f returns true with a reference to it as a parameter or None if no such item exists in the queue.

fn pop_if<P>(&mut self, predicate: P) -> Option<T> where P: Fn(&T) -> bool

Returns the most recently pushed block if f returns true with a reference to it as a parameter, otherwise None. Will not destroy a block if a reference to it has previously been returned by use_last_ref, but rather clone it.