Module ethcore_util::bytes

Unified interfaces for bytes operations on basic types

Examples

extern crate ethcore_util as util;

fn bytes_convertable() {
    use util::bytes::BytesConvertable;

    let arr = [0; 5];
    let slice: &[u8] = arr.bytes();
}

fn main() {
    bytes_convertable();
}

Structs

PrettySlice
Slie pretty print helper

Enums

BytesRef
A byte collection reference that can either be a slice or a vector

FromBytesError
Bytes array deserialization error

Traits

BytesConvertable
Slice of bytes to underlying memory

FromRawBytes
Value that can be serialized from bytes array

FromRawBytesVariable
Value that can be serialized from variable-length byte array

Populatable
Simple trait to allow for raw population of a Sized object from a byte slice.

ToBytesWithMap
Value that serializes directly to variable-sized byte array and stores map

ToPretty
Trait to allow a type to be pretty-printed in format!, where unoverridable defaults cannot otherwise be avoided.

VecLike
Vector like object

Type Definitions

Bytes
Vector of bytes