Struct ethcore_util::memorydb::MemoryDB

pub struct MemoryDB { /* fields omitted */ }

Reference-counted memory-based HashDB implementation.

Use new() to create a new database. Insert items with insert(), remove items with remove(), check for existence with containce() and lookup a hash to derive the data with get(). Clear with clear() and purge the portions of the data that have no references with purge().

Example

extern crate ethcore_util;
use ethcore_util::hashdb::*;
use ethcore_util::memorydb::*;
fn main() {
  let mut m = MemoryDB::new();
  let d = "Hello world!".as_bytes();

  let k = m.insert(d);
  assert!(m.contains(&k));
  assert_eq!(m.get(&k).unwrap(), d);

  m.insert(d);
  assert!(m.contains(&k));

  m.remove(&k);
  assert!(m.contains(&k));

  m.remove(&k);
  assert!(!m.contains(&k));

  m.remove(&k);
  assert!(!m.contains(&k));

  m.insert(d);
  assert!(!m.contains(&k));
  m.insert(d);
  assert!(m.contains(&k));
  assert_eq!(m.get(&k).unwrap(), d);

  m.remove(&k);
  assert!(!m.contains(&k));
}

Methods

impl MemoryDB

fn new() -> MemoryDB

Create a new instance of the memory DB.

fn clear(&mut self)

Clear all data from the database.

Examples

extern crate ethcore_util;
use ethcore_util::hashdb::*;
use ethcore_util::memorydb::*;
fn main() {
  let mut m = MemoryDB::new();
  let hello_bytes = "Hello world!".as_bytes();
  let hash = m.insert(hello_bytes);
  assert!(m.contains(&hash));
  m.clear();
  assert!(!m.contains(&hash));
}

fn purge(&mut self)

Purge all zero-referenced data from the database.

fn drain(&mut self) -> H256FastMap<(DBValue, i32)>

Return the internal map of hashes to data, clearing the current state.

fn raw(&self, key: &H256) -> Option<(DBValue, i32)>

Grab the raw information associated with a key. Returns None if the key doesn't exist.

Even when Some is returned, the data is only guaranteed to be useful when the refs > 0.

fn mem_used(&self) -> usize

Returns the size of allocated heap memory

fn remove_and_purge(&mut self, key: &H256)

Remove an element and delete it from storage if reference count reaches zero.

fn consolidate(&mut self, other: Self)

Consolidate all the entries of other into self.