Module ethcore_util::network

Network and general IO module.

Example usage for craeting a network service and adding an IO handler:

extern crate ethcore_util as util;
use util::*;

struct MyHandler;

#[derive(Clone)]
struct MyMessage {
    data: u32
}

impl NetworkProtocolHandler<MyMessage> for MyHandler {
    fn initialize(&self, io: &NetworkContext<MyMessage>) {
        io.register_timer(0, 1000);
    }

    fn read(&self, io: &NetworkContext<MyMessage>, peer: &PeerId, packet_id: u8, data: &[u8]) {
        println!("Received {} ({} bytes) from {}", packet_id, data.len(), peer);
    }

    fn connected(&self, io: &NetworkContext<MyMessage>, peer: &PeerId) {
        println!("Connected {}", peer);
    }

    fn disconnected(&self, io: &NetworkContext<MyMessage>, peer: &PeerId) {
        println!("Disconnected {}", peer);
    }

    fn timeout(&self, io: &NetworkContext<MyMessage>, timer: TimerToken) {
        println!("Timeout {}", timer);
    }

    fn message(&self, io: &NetworkContext<MyMessage>, message: &MyMessage) {
        println!("Message {}", message.data);
    }
}

fn main () {
    let mut service = NetworkService::<MyMessage>::new(NetworkConfiguration::new_local()).expect("Error creating network service");
    service.register_protocol(Arc::new(MyHandler), "myproto", &[1u8]);
    service.start().expect("Error starting service");

    // Wait for quit condition
    // ...
    // Drop the service
}

Reexports

pub use network::host::NetworkIoMessage::User as UserMessage;

Structs

NetworkConfiguration
Network service configuration

NetworkContext
IO access point. This is passed to all IO handlers and provides an interface to the IO subsystem.

NetworkService
IO Service with networking Message defines a notification data type.

NetworkStats
Network statistics structure

Enums

NetworkError
Network error.

NetworkIoMessage
Messages used to communitate with the event loop from other threads.

NonReservedPeerMode
Non-reserved peer modes.

Traits

NetworkProtocolHandler
Network IO protocol handler. This needs to be implemented for each new subprotocol. All the handler function are called from within IO event loop. Message is the type for message data.

Functions

is_valid_node_url
Check if node url is valid

Type Definitions

PacketId
Protocol handler level packet id

PeerId
Local (temporary) peer session ID.