Struct ethcore_util::overlaydb::OverlayDB

pub struct OverlayDB { /* fields omitted */ }

Implementation of the HashDB trait for a disk-backed database with a memory overlay.

The operations insert() and remove() take place on the memory overlay; batches of such operations may be flushed to the disk-backed DB with commit() or discarded with revert().

lookup() and contains() maintain normal behaviour - all insert() and remove() queries have an immediate effect in terms of these functions.

Methods

impl OverlayDB

fn new(backing: Arc<Database>, col: Option<u32>) -> OverlayDB

Create a new instance of OverlayDB given a backing database.

fn commit_to_batch(&mut self,
batch: &mut DBTransaction)
-> Result<u32, UtilError>

Commit all operations to given batch.

fn revert(&mut self)

Revert all operations on this object (i.e. insert()s and remove()s) since the last commit().

fn commit_refs(&self, key: &H256) -> i32

Get the number of references that would be committed.