Enum ethcore_util::trie::node::Node

pub enum Node {
    Empty,
    Leaf(NodeKey, DBValue),
    Extension(NodeKey, DBValue),
    Branch([NodeKey; 16], Option<DBValue>),
}

Type of node in the trie and essential information thereof.

Variants
Empty

Null trie node; could be an empty root or an empty branch entry.

Leaf(NodeKey, DBValue)

Leaf node; has key slice and value. Value may not be empty.

Extension(NodeKey, DBValue)

Extension node; has key slice and node data. Data may not be null.

Branch([NodeKey; 16], Option<DBValue>)

Branch node; has array of 16 child nodes (each possibly null) and an optional immediate node data.

Methods

impl Node

fn decoded(node_rlp: &[u8]) -> Node

Decode the node_rlp and return the Node.

fn encoded(&self) -> Bytes

Encode the node into RLP.

Will always return the direct node RLP even if it's 32 or more bytes. To get the RLP which would be valid for using in another node, use encoded_and_added().