Struct ethcore_util::trie::sectriedb::SecTrieDB

pub struct SecTrieDB<'db> { /* fields omitted */ }

A Trie implementation which hashes keys and uses a generic HashDB backing database.

Use it as a Trie trait object. You can use raw() to get the backing TrieDB object.

Methods

impl<'db> SecTrieDB<'db>

fn new(db: &'db HashDB, root: &'db H256) -> Result<Self>

Create a new trie with the backing database db and empty root

Initialise to the state entailed by the genesis block. This guarantees the trie is built correctly. Returns an error if root does not exist.

fn raw(&self) -> &TrieDB

Get a reference to the underlying raw TrieDB struct.

fn raw_mut(&mut self) -> &mut TrieDB<'db>

Get a mutable reference to the underlying raw TrieDB struct.